Update README.md
This model aims to assist medical professionals in accurately diagnosing and predicting the presence of cancer in patients.

This model takes in diverse input data, which can include medical images (such as X-rays, MRIs, or CT scans), patient medical histories, and potentially genetic information. The model then employs complex mathematical computations and pattern recognition methods to extract meaningful features and relationships from this data.

This model was exposed to large datasets of both cancer-positive and healthy cases, allowing it to learn and discern subtle distinctions between the two categories. As the model learns, it refines its ability to differentiate between benign and malignant characteristics in the data.

This model has been trained and can be used for real-world applications. When presented with new, unseen data, the cancer detector model can make predictions with varying levels of confidence about whether a patient is likely to have cancer. This model's predictions can then be used as an additional tool for healthcare practitioners to consider when making diagnoses and treatment decisions.